chore(modals): rename ModalWrapper to Modal
