+++ client/src/app/primitives/index.js
@@ -15,5 +15,5 @@ export { default as Loader } from './Loader';
 export { default as TabLinks } from './TabLinks';
 export { default as TabContainer } from './TabContainer';
 export { default as Tab } from './Tab';
-export { ModalWrapper } from './modal';
+export { Modal } from './modal';
 export { ToastContainer } from './toast';
