+++ client/src/app/primitives/modal/index.js
@@ -8,4 +8,4 @@
  * except in compliance with the MIT License.
  */
 
-export { default as ModalWrapper } from './ModalWrapper';
+export { default as Modal } from './Modal';
